requirements.txt
faiss-cpu
groq
langchain-grok
PyPDF2
langchain_google_genai
langchain
streamlit
langchain_community
python-dotenv
pypdf